requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
feedparser==6.0.10
beautifulsoup4==4.12.2
aiohttp==3.9.1
pydantic==2.5.2
praw==7.7.1
requests==2.31.0
python-dateutil==2.8.2
pyyaml==6.0.1
openai==1.12.0 
httpx==0.27.2
beautifulsoup4==4.12.2
croniter==2.0.2 